Modelo actualizado pendiente de revisión

+++ src/main/java/es/agrobook/api/model/explotacion/ExplotacionClasificacion.java
@@ -13,7 +13,7 @@ import lombok.*;
 public class ExplotacionClasificacion {
 	
 	@Id
-	private long id;
+	private byte id;
 
 	@Column(nullable = false)
 	private String descripcion;
