+++ src/main/java/es/agrobook/api/model/persona/PersonaTipoAsociacion.java
@@ -11,7 +11,7 @@ import lombok.*;
 public class PersonaTipoAsociacion {
 	
 	@Id
-	private long id;
+	private byte id;
 
 	@Column(nullable = false)
 	private String descripcion;
